requirements.txt
pytest
selenium
requests
allure-pytest
webdriver-manager

pytest-html
pytest-xdist